agent-automation/requirements.txt
PyYAML>=6.0
watchdog>=3.0.0

